♻️ refactor: Split download main logic into `download_manager.py`
